added group members

+++ src/groups/groups.service.ts
@@ -5,6 +5,8 @@ import { Repository } from 'typeorm';
 import { createGroupDTO } from './dto/create-group.dto';
 import { User } from 'src/entities/user.entity';
 import { GroupMember } from 'src/entities/group-member.entity';
+import { createGroupMemberDTO } from './dto/create-groupMember.dto';
+import { deleteGroupMemberDTO } from './dto/delete-groupMember.dto';
 
 @Injectable()
 export class GroupsService {
@@ -39,17 +41,59 @@ export class GroupsService {
     }
 
     async deleteGroup(groupId:number){
+        const group=await this.groupRepo.findOneBy({id:groupId})
+        if(!group){
+            throw new HttpException('invalid group id',404)
+        }
         const result=await this.groupRepo.delete({id:groupId})
         return result
     }
 
     // GROUP-MEMBER CRUD
 
-    async addGroupMember(){
+    async addGroupMember(createGroupMemberDTO:createGroupMemberDTO){
+        const group=await this.groupRepo.findOneBy({id:createGroupMemberDTO.groupId})
+        if(!group){
+            throw new HttpException('invalid group id',404)
+        }
+        const user=await this.userRepo.findOneBy({id:createGroupMemberDTO.userId})
+        
+        if(!user){
+            throw new HttpException('Invalid User',404)
+        }
 
+        const groupMember=await this.groupMemberRepo.findOneBy({groupId:{id:createGroupMemberDTO.groupId},userId:{id:createGroupMemberDTO.userId}})
+        
+        if(groupMember && groupMember.isMember==true){
+            throw new HttpException('user already member of the group',404)
+        }
+
+        const newGroupMember=await this.groupMemberRepo.create({
+            groupId:group,
+            userId:user
+        })
+
+        return await this.groupMemberRepo.save(newGroupMember)
+        
     }
 
-    async removeGroupMember(){
+    async removeGroupMember(deleteGroupMemberDTO:deleteGroupMemberDTO){
+        const group=await this.groupRepo.findOneBy({id:deleteGroupMemberDTO.groupId})
+        if(!group){
+            throw new HttpException('invalid group id',404)
+        }
+        const user=await this.userRepo.findOneBy({id:deleteGroupMemberDTO.userId})
         
+        if(!user){
+            throw new HttpException('Invalid User',404)
+        }
+
+        const groupMember=await this.groupMemberRepo.findOneBy({groupId:{id:deleteGroupMemberDTO.groupId},userId:{id:deleteGroupMemberDTO.userId}})
+
+        if(groupMember.isMember==false){
+            throw new HttpException('user not a member of the group',404)
+        }
+
+        return await this.groupMemberRepo.update(groupMember.id,{isMember:false})
     }
 }
